tests: reorganise test folder structure
