+++ tests/resources/test_resource_configuration.py
