+++ tests/resources/test_serialization_integration.py
